requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cffi==1.16.0
contextlib2==21.6.0
cryptography==42.0.8
Django>=4.2
django-cors-headers==4.4.0
django-ninja==1.2.1
django-ninja-extra==0.21.1
django-ninja-jwt==5.3.2
gunicorn==22.0.0
iniconfig==2.0.0
injector==0.22.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
